requirements.txt
streamlit
assemblyai
websocket-client
sounddevice